pyproject.toml
[project]
name = "ai-receptionist-for-hostpital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "livekit-agents[cartesia,deepgram,groq,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "motor>=3.7.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "twilio>=9.10.0",
    "uvicorn>=0.40.0",
]
